+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - [cider#3792](https://github.com/clojure-emacs/cider/issues/3792): Don't fail
   if context macroexpansion throws an exception.
+- [#122](https://github.com/alexander-yakushev/compliment/pull/122): Introduce
+  sorting based on candidate priorities.
 - [#121](https://github.com/alexander-yakushev/compliment/pull/121): Remove
   deprecated `:plain-candidates` flag.