+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### master (unreleased)
 
+### 0.7.0 (2025-03-25)
+
 - [cider#3792](https://github.com/clojure-emacs/cider/issues/3792): Don't fail
   if context macroexpansion throws an exception.
 - [#122](https://github.com/alexander-yakushev/compliment/pull/122): Introduce